refactor: rename install command to setup
paso install -> paso setup. We are not installing opencode or claude,
we are setting them up with paso skills and commands. Rename the
package, files, exported func, cobra Use/Short, flag help text,
and all test references to match.
