+++ internal/cli/setup/commands/paso-delegate.md
