+++ internal/cli/setup/skills/paso/SKILL.md
