+++ internal/cli/setup/utils.go
@@ -1,4 +1,4 @@
-package install
+package setup
 
 import (
 	"fmt"
