+++ internal/testing/snapshots/cli/testdata/help/paso-setup.golden
@@ -1,0 +1,10 @@
+Set up paso skills and commands for AI tools
+
+Usage:
+  paso setup [command]
+
+Available Commands:
+  claude      Set up paso for Claude Code (~/.claude/)
+  opencode    Set up paso for OpenCode (~/.config/opencode/)
+
+Use "paso setup [command] --help" for more information about a command.
